Move grunt tasks to custom dir
