+++ grunt/custom/github-pre-release.js
@@ -1,6 +1,6 @@
-const getUserInput = require( "./tools/get-user-input" );
-const githubApi = require( "./tools/github-api" );
-const uploadToGitHub = require( "./tools/upload-to-github" );
+const getUserInput = require( "../config/tools/get-user-input" );
+const githubApi = require( "../config/tools/github-api" );
+const uploadToGitHub = require( "../config/tools/upload-to-github" );
 
 /**
  * Throws an error.
